feat: move loader and text to ui directory
