+++ src/components/molecules/PairDetail.tsx
@@ -7,7 +7,7 @@ import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 import TreeItem from '@mui/lab/TreeItem';
 import { AddButton } from "../ui/buttons/AddButton";
-import { WarningText } from "../atoms/texts";
+import { WarningText } from "../ui/texts";
 import NotificationImportantIcon from '@mui/icons-material/NotificationImportant';
 import { RemoveButton } from "../ui/buttons/RemoveButton";
 import { AddWhereCondModal } from "./AddWhereCondModal";
