+++ src/components/molecules/action-settings/scrapeSchema.tsx
@@ -1,5 +1,5 @@
 import React, { forwardRef, useCallback, useImperativeHandle, useRef } from 'react';
-import { WarningText } from "../../atoms/texts";
+import { WarningText } from "../../ui/texts";
 import InfoIcon from "@mui/icons-material/Info";
 import { KeyValueForm } from "../KeyValueForm";
